Monolith: ensure session based authentication works
