+++ app/controllers/ui/application_controller.rb
@@ -1,0 +1,5 @@
+module Ui
+  class ApplicationController < ActionController::Base
+    include SharedControllerMethods
+  end
+end
