+++ app/controllers/ui/login_test_controller.rb
@@ -1,0 +1,6 @@
+module Ui
+  class LoginTestController < ApplicationController
+    def show
+    end
+  end
+end
