+++ app/controllers/v0/sessions_controller.rb
@@ -8,8 +8,10 @@ module V0
       authorize user, :show?
       if user && user.authenticate_with_legacy_support(params[:password])
         # $analytics.track("login:successful", user.id)
+        session[:user_id] = user.id
         render json: { access_token: user.access_token!.token }, status: :ok
       else
+        session[:user_id] = nil
         raise Smartcitizen::UnprocessableEntity.new({
           message: {password: 'is incorrect'}, # to be removed
           password: 'is incorrect' # to replace the above
